third_party/blink/web_tests/http/tests/subresource_filter/image-allow-disallow-transitions-expected.txt

CONSOLE ERROR: Chrome blocked resource http://127.0.0.1:8000/subresource_filter/resources/gamma.png on this site because this site tends to show ads that interrupt, distract, mislead, or prevent user control. Learn more at https://www.chromestatus.com/feature/5738264052891648
CONSOLE ERROR: Chrome blocked resource http://127.0.0.1:8000/subresource_filter/resources/gamma.png on this site because this site tends to show ads that interrupt, distract, mislead, or prevent user control. Learn more at https://www.chromestatus.com/feature/5738264052891648
CONSOLE ERROR: Chrome blocked resource http://127.0.0.1:8000/subresource_filter/resources/gamma.png on this site because this site tends to show ads that interrupt, distract, mislead, or prevent user control. Learn more at https://www.chromestatus.com/feature/5738264052891648
CONSOLE ERROR: Chrome blocked resource http://127.0.0.1:8000/subresource_filter/resources/delta.png on this site because this site tends to show ads that interrupt, distract, mislead, or prevent user control. Learn more at https://www.chromestatus.com/feature/5738264052891648
CONSOLE ERROR: Chrome blocked resource http://127.0.0.1:8000/subresource_filter/resources/delta.png on this site because this site tends to show ads that interrupt, distract, mislead, or prevent user control. Learn more at https://www.chromestatus.com/feature/5738264052891648
CONSOLE ERROR: Chrome blocked resource http://127.0.0.1:8000/subresource_filter/resources/delta.png on this site because this site tends to show ads that interrupt, distract, mislead, or prevent user control. Learn more at https://www.chromestatus.com/feature/5738264052891648
This is a testharness.js-based test.
[FAIL] State transition primary to fallback. 
  assert_equals: Image has incorrect width for the expected final state. expected 0 but got 16
[FAIL] State transition fallback to fallback. 
  assert_equals: Image has incorrect width for the expected final state. expected 0 but got 16
[FAIL] State transition collapsed to fallback. 
  assert_equals: Image has incorrect width for the expected final state. expected 0 but got 16
Harness: the test ran to completion.
